+++ test-workspace/config/vue2-ssr/server.js
@@ -6,7 +6,7 @@ const balmrc = require('../balmrc');
 
 const scripts = Object.assign(base, {
   entry: {
-    server: './vue-ssr/app/scripts/entry-server.js'
+    server: './vue2-ssr/app/scripts/entry-server.js'
   },
   // This allows webpack to handle dynamic imports in a Node-appropriate
   // fashion, and also tells `vue-loader` to emit server-oriented code when
@@ -29,7 +29,7 @@ const scripts = Object.assign(base, {
 
 const balmConfig = Object.assign(balmrc, {
   roots: {
-    source: 'vue-ssr/app'
+    source: 'vue2-ssr/app'
   },
   scripts
 });