+++ test-workspace/vue2/scripts/main.js
@@ -1,0 +1,14 @@
+import Vue from 'vue';
+import App from '@/views/layouts/app';
+import BalmUI from 'balm-ui';
+// import 'balm-ui/dist/balm-ui.css';
+
+Vue.config.productionTip = false;
+
+Vue.use(BalmUI);
+
+new Vue({
+  el: '#app',
+  components: { App },
+  template: '<app/>'
+});